Scrapes from Carmichael's dining hall!
Scrapes from Carm's dining hall: takes the menu for each day and saves it into plain text format. Then, create a list object of this text so that students can search "(a certain food)" and be able to find out whether or not it is being served in the dining halls.